fixes local var scope in handling proxy whitelist lifecycle

+++ src/main.py
@@ -19,6 +19,9 @@ load_dotenv()
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     session = await get_global_session()
+    api_url = None
+    proxy_header_token = None
+    payload = None
     if os.getenv('PROXY_TOKEN') and os.getenv('USE_PROXY', 'False') == 'True':
         async with session.get("https://api.ipify.org/") as ip_response:
             ip = await ip_response.text()
@@ -30,7 +33,8 @@ async def lifespan(app: FastAPI):
             payload = {"ip": ip}
             await session.post(api_url, headers=proxy_header_token, json=payload)
     yield
-    await session.delete(api_url, headers=proxy_header_token, json=payload)
+    if api_url and proxy_header_token and payload:
+        await session.delete(api_url, headers=proxy_header_token, json=payload)
     await close_global_session()
     await r.close()
